correção upload

+++ modulos/painel/controles/usuariosControle.php
@@ -39,14 +39,14 @@ class usuariosControle extends painelControle
         $this->_view->assign('lista', $lista);
     }
 
-    
+
     public function index()
     {
         $this->_view->assign('titulo', 'Gerenciamnto de usuários');
 
         $this->_view->addNavLink('usuarios/registro', 'Gerenciamnto de usuários');
-        
-        $this->_view->assign('FormAction', BASE_URL.'painel/usuarios/novo');
+
+        $this->_view->assign('FormAction', BASE_URL . 'painel/usuarios/novo');
         $this->_view->addConteudo('usuarioCadastro');
         $this->_view->renderizar();
     }
@@ -55,14 +55,14 @@ class usuariosControle extends painelControle
     {
         $post = $this->POST();
         unset($post['confirma_senha']);
-        
+
         if ($this->POST()) {
             $this->_user->_load($post);
 
             if (isset($post['USU_imagem'])) {
                 $this->_user->upImagem();
             }
-            
+
             if ($this->_user->_grava()) {
                 Sessao::addMsg('sucesso', 'Usuário gravado com Sucesso');
             } else {
@@ -71,22 +71,24 @@ class usuariosControle extends painelControle
         }
         $this->index();
     }
-    
-    
+
+
     public function editar($id)
     {
         if (is_numeric($id)) {
             $user = $this->_user->_selectUser('USU_id', $id);
             if (is_array($user)) {
-                unset($user[0]['USU_senha']);// deleta o campo senha para não ser exibido pela consulta;
+                unset($user[0]['USU_senha']); // deleta o campo senha para não ser exibido pela consulta;
                 $this->_view->assign('campos', json_encode($user[0]));
             }
-            
-            
+
             if ($this->POST()) {
+
                 $this->_user->_load($this->POST());
                 if (!empty($this->POST('USU_imagem'))) {
-                    unlink(RAIZ.'upload/'.$user[0]['USU_imagem']);
+                    if (is_readable(RAIZ . 'upload/' . $user[0]['USU_imagem'])) {
+                        unlink(RAIZ . 'upload/' . $user[0]['USU_imagem']);
+                    }
                     $this->_user->upImagem();
                 }
                 if ($this->_user->_atualiza($id)) {
@@ -99,13 +101,13 @@ class usuariosControle extends painelControle
         }
 
         $this->_view->addNavLink('painel/usuarios/editar', 'Editar usuário');
-        $this->_view->assign('titulo', 'Editando usuário:'.$user[0]['USU_nome']);
-        $this->_view->assign('FormAction', BASE_URL.'painel/usuarios/editar/'.$user[0]['USU_id']);
+        $this->_view->assign('titulo', 'Editando usuário:' . $user[0]['USU_nome']);
+        $this->_view->assign('FormAction', BASE_URL . 'painel/usuarios/editar/' . $user[0]['USU_id']);
         $this->_view->assign('usuario', $user[0]);
         $this->_view->addConteudo('usuarioCadastro');
         $this->_view->renderizar();
     }
-    
+
     public function excluir($id = false)
     {
         if (is_numeric($id)) {
@@ -115,7 +117,7 @@ class usuariosControle extends painelControle
                 Sessao::addMsg('erro', 'Usuário não pode ser excluído');
             }
         } elseif ($this->POST()) {
-            if ($this->_db->_query('DELETE FROM usuarios WHERE USU_id IN ('.$this->POST('selUser').')')) {
+            if ($this->_db->_query('DELETE FROM usuarios WHERE USU_id IN (' . $this->POST('selUser') . ')')) {
                 Sessao::addMsg('sucesso', 'Todos os usuários selecionados foram excluidos.');
             } else {
                 Sessao::addMsg('erro', 'Os usuários não puderam ser excluídos');
